+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Bug Fixes
+* 
+* [#156](https://github.com/babylonlabs-io/covenant-emulator/pull/156) fix: overflow in parse hardened
+
 ## v0.16.0-rc.2
 
 ### Improvements